# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Electronic Document Management System API - Backend service for document storage, search, and management\nSupports document ingestion, full-text search, word cloud generation, and file system browsing",
        "title": "godocs Backend API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/drummonds/godocs"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/clean": {
            "post": {
                "description": "Remove database entries for missing files and move orphaned files to ingress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Clean database",
                "responses": {
                    "200": {
                        "description": "Job created with jobId",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/document/upload": {
            "post": {
                "description": "Upload a new document file to the ingress folder for processing",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Upload path (relative to ingress folder)",
                        "name": "path",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Document file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Path to uploaded file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/document/{id}": {
            "get": {
                "description": "Retrieve document details by ULID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Get a document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ULID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document details",
                        "schema": {
                            "$ref": "#/definitions/database.Document"
                        }
                    },
                    "404": {
                        "description": "Document not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/documents/filesystem": {
            "get": {
                "description": "Retrieve the complete document folder structure as a tree",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Get document filesystem tree",
                "responses": {
                    "200": {
                        "description": "Complete filesystem tree",
                        "schema": {
                            "$ref": "#/definitions/engine.fullFileSystem"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/documents/latest": {
            "get": {
                "description": "Retrieve the most recently ingested documents with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Get latest documents",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated documents with metadata",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/folder/{folder}": {
            "get": {
                "description": "Retrieve all documents in a specific folder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folders"
                ],
                "summary": "Get folder contents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder name",
                        "name": "folder",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of documents in folder",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Document"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/jobs": {
            "get": {
                "description": "Retrieve a list of recent jobs with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get recent jobs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of jobs to return (default: 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of jobs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Job"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/jobs/active": {
            "get": {
                "description": "Retrieve all jobs that are currently running or pending",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get active jobs",
                "responses": {
                    "200": {
                        "description": "List of active jobs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Job"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "description": "Retrieve details of a specific job by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID (ULID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details",
                        "schema": {
                            "$ref": "#/definitions/database.Job"
                        }
                    },
                    "400": {
                        "description": "Invalid job ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "Search all documents using PostgreSQL full-text search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search documents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "term",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "schema": {
                            "$ref": "#/definitions/engine.fullFileSystem"
                        }
                    },
                    "204": {
                        "description": "No results found"
                    },
                    "404": {
                        "description": "Empty search term",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/search/reindex": {
            "post": {
                "description": "Rebuild the full-text search index for all documents",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Reindex search documents",
                "responses": {
                    "200": {
                        "description": "Reindex successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Reindex failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/wordcloud": {
            "get": {
                "description": "Retrieve the top N most frequent words from all documents for word cloud visualization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WordCloud"
                ],
                "summary": "Get word cloud data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of words to return (default: 100, max: 500)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Word cloud data with words, metadata, and count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "database.Document": {
            "type": "object",
            "properties": {
                "documentType": {
                    "description": "type of document (pdf, txt, etc)",
                    "type": "string"
                },
                "folder": {
                    "type": "string"
                },
                "fullText": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "ingressTime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "description": "full path to the file",
                    "type": "string"
                },
                "stormID": {
                    "description": "ID field (kept as StormID for backward compatibility)",
                    "type": "integer"
                },
                "ulid": {
                    "description": "Have a smaller (than hash) id that can be used in URL's, hopefully speed things up",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "database.Job": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentStep": {
                    "description": "Human-readable current step",
                    "type": "string"
                },
                "error": {
                    "description": "Error message if failed",
                    "type": "string"
                },
                "id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "message": {
                    "description": "Status message",
                    "type": "string"
                },
                "progress": {
                    "description": "0-100",
                    "type": "integer"
                },
                "result": {
                    "description": "JSON result data",
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/database.JobStatus"
                },
                "totalSteps": {
                    "description": "Total number of steps",
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/database.JobType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "database.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "JobStatusPending",
                "JobStatusRunning",
                "JobStatusCompleted",
                "JobStatusFailed",
                "JobStatusCancelled"
            ]
        },
        "database.JobType": {
            "type": "string",
            "enum": [
                "ingestion",
                "cleanup",
                "wordcloud",
                "search_reindex"
            ],
            "x-enum-varnames": [
                "JobTypeIngestion",
                "JobTypeCleanup",
                "JobTypeWordCloud",
                "JobTypeSearchReindex"
            ]
        },
        "engine.fileTreeStruct": {
            "type": "object",
            "properties": {
                "childrenIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fileURL": {
                    "type": "string"
                },
                "fullPath": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isDir": {
                    "type": "boolean"
                },
                "modDate": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "openable": {
                    "type": "boolean"
                },
                "parentID": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "ulid": {
                    "type": "string"
                }
            }
        },
        "engine.fullFileSystem": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fileSystem": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/engine.fileTreeStruct"
                    }
                }
            }
        }
    },
...
}